# health_pal_backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  api_types.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  api_types.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  handlers.RecordSittingTimeRequest:
    properties:
      date:
        type: string
      duration_minutes:
        minimum: 0
        type: integer
    required:
    - date
    - duration_minutes
    type: object
...
  models.DailyNutritionSummary:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      micronutrients_json:
        description: Storing as JSON string
        type: string
      record_date:
        description: Unique per user per date
        type: string
      total_calories:
        type: number
      total_carbohydrates:
        type: number
      total_fats:
        type: number
      total_protein:
        type: number
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.DailyNutritionSummaryResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      micronutrients_json:
        type: string
      record_date:
        type: string
      total_calories:
        type: number
      total_carbohydrates:
        type: number
      total_fats:
        type: number
      total_protein:
        type: number
      user_id:
        type: integer
    type: object
...
  models.ReminderResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      reminder_text:
        type: string
      scheduled_time:
        type: string
      sent_at:
        type: string
      status:
        type: string
      user_id:
        type: integer
    type: object
  models.SittingTimeResponse:
    properties:
      created_at:
        type: string
      date:
        type: string
      duration_minutes:
        type: integer
      id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API documentation for the Health Pal backend.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Health Pal Backend API
  version: "1.0"
paths:
...
  /api/v1/nutrition/daily-summary:
    get:
      description: Retrieves the aggregated daily nutrition data for the authenticated
        user for a given date.
      parameters:
      - description: Date in YYYY-MM-DD format
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Daily nutrition summary retrieved successfully
          schema:
            $ref: '#/definitions/models.DailyNutritionSummaryResponse'
        "400":
          description: Invalid date format
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
        "404":
          description: Nutrition summary not found for the given date
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get daily nutrition summary
      tags:
      - Nutrition
  /api/v1/nutrition/manual-entry:
    post:
      consumes:
      - application/json
      description: Allows users to manually input nutrition data for a specific date.
      parameters:
      - description: Nutrition data to log
        in: body
        name: nutrition
        required: true
        schema:
          $ref: '#/definitions/models.DailyNutritionSummary'
      produces:
      - application/json
      responses:
        "201":
          description: Nutrition data logged successfully
          schema:
            $ref: '#/definitions/api_types.SuccessResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Manually log nutrition data
      tags:
      - Nutrition
  /api/v1/reminders:
    get:
      description: Retrieves all scheduled and sent reminders for the authenticated
        user.
      produces:
      - application/json
      responses:
        "200":
          description: Reminders retrieved successfully
          schema:
            items:
              $ref: '#/definitions/models.ReminderResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get reminders
      tags:
      - Reminders
  /api/v1/reminders/schedule:
    post:
      consumes:
      - application/json
      description: Schedules a personalized reminder based on user data and health
        goals.
      parameters:
      - description: Reminder details (e.g., scheduled_time, type)
        in: body
        name: reminder
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Reminder scheduled successfully
          schema:
            $ref: '#/definitions/api_types.SuccessResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Schedule a reminder
      tags:
      - Reminders
  /api/v1/sitting-times:
    post:
      consumes:
      - application/json
      description: Records or updates the daily sitting time in minutes for the authenticated
        user.
      parameters:
      - description: Sitting time data
        in: body
        name: sittingTime
        required: true
        schema:
          $ref: '#/definitions/handlers.RecordSittingTimeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Sitting time recorded successfully
          schema:
            $ref: '#/definitions/api_types.SuccessResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Record daily sitting time
      tags:
      - SittingTime
  /api/v1/sitting-times/history:
    get:
      description: Retrieves all recorded sitting times in minutes for the authenticated
        user.
      produces:
      - application/json
      responses:
        "200":
          description: List of sitting time entries
          schema:
            items:
              $ref: '#/definitions/models.SittingTimeResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get sitting time history
      tags:
      - SittingTime
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"